+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 - Add imperial support for UDCF import
+- Desktop: combo box with fresh, sea water and EN 13319 in dive planner
+- Desktop: Changed "salinity" to "water type" at dive planner and dive info
 # add new entries above this line